+++ app/scripts/loadout/dimLoadout.directive.js
@@ -212,7 +212,7 @@
               .values()
               .flatten()
               .findWhere({
-                sort: item.sort,
+                sort: item.bucket.sort,
                 tier: dimItemTier.exotic,
                 equipped: true
               })